課題１１終了

+++ 課題11/Enemy.h
@@ -3,7 +3,7 @@
 class Enemy
 {
 	int hp;
-	float atk;
+	int atk;
 	int def;
 
 public:
